requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.23
datasets==1.16.1
decorator==5.1.0
dill==0.3.4
discord.py==1.7.3
dnspython==2.1.0
fast-ml==3.68
filelock==3.4.0
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2021.11.1
fuzzywuzzy==0.18.0
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h11==0.12.0
h5py==2.10.0
huggingface-hub==0.2.1
idna==2.10
ipython-genutils==0.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-core==4.7.1
Keras==2.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.3
motor==2.5.1
multidict==5.2.0
multiprocess==0.70.12.2
nnfs==0.5.1
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.23
protobuf==3.15.6
py-cord @ git+https://github.com/Pycord-Development/pycord@125db3ab98eecc4985970f5082cca427c79278a0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.1
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
resnet==0.1
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.6.1
seaborn==0.11.1
selenium==4.0.0
six==1.15.0
sklearn==0.0
slid==0.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.5.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.12.5
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.0.0
xxhash==2.0.2
yarl==1.7.2
zipp==3.4.1